requirements.txt
PyMuPDF
pandas
numpy
openpyxl
python-dotenv
python-dateutil
google-api-python-client 
google-auth-httplib2
google-auth-oauthlib 
streamlit>=1.40.0
altair<5
standard-imghdr

google-genai
openai
pillow
groq
supabase

